- All initiative.py slash commands have been migrated to the new models
