+++ D4e/d4e_functions.py
@@ -381,102 +381,6 @@ async def gm_check(ctx, engine, guild=None):
         return False
 
 
-async def edit_stats(ctx, engine, name: str, bot):
-    async_session = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-    try:
-        if engine == None:
-            engine = get_asyncio_db_engine(user=USERNAME, password=PASSWORD, host=HOSTNAME, port=PORT, db=SERVER_DATA)
-        guild = await get_guild(ctx, None)
-
-        Character_Model = await get_character(name, ctx, guild=guild, engine=engine)
-        condition_dict = {}
-        for con in await Character_Model.conditions():
-            await asyncio.sleep(0)
-            condition_dict[con.title] = con.number
-        editModal = D4eEditCharacterModal(
-            character=await Character_Model.character(), cons=condition_dict, ctx=ctx, engine=engine, title=name, bot=bot
-        )
-        await ctx.send_modal(editModal)
-
-        return True
-
-    except Exception:
-        return False
-
-
-# D&D 4e Specific
-class D4eEditCharacterModal(discord.ui.Modal):
-    def __init__(self, character, cons: dict, ctx: discord.ApplicationContext, engine, bot, *args, **kwargs):
-        self.character = character
-        self.cons = (cons,)
-        self.name = character.name
-        self.player = ctx.user.id
-        self.ctx = ctx
-        self.engine = engine
-        self.bot = bot
-        super().__init__(
-            discord.ui.InputText(label="AC", placeholder="Armor Class", value=cons["AC"]),
-            discord.ui.InputText(label="Fort", placeholder="Fortitude", value=cons["Fort"]),
-            discord.ui.InputText(label="Reflex", placeholder="Reflex", value=cons["Reflex"]),
-            discord.ui.InputText(label="Will", placeholder="Will", value=cons["Will"]),
-            *args,
-            **kwargs,
-        )
-
-    async def callback(self, interaction: discord.Interaction):
-        self.stop()
-        await interaction.response.send_message(f"{self.name} Updated")
-        async_session = sessionmaker(self.engine, expire_on_commit=False, class_=AsyncSession)
-        async with async_session() as session:
-            result = await session.execute(
-                select(Global).where(
-                    or_(
-                        Global.tracker_channel == self.ctx.interaction.channel_id,
-                        Global.gm_tracker_channel == self.ctx.interaction.channel_id,
-                    )
-                )
-            )
-            guild = result.scalars().one()
-
-        discord.Embed(
-            title="Character Updated (D&D 4e)",
-            fields=[
-                discord.EmbedField(name="Name: ", value=self.name, inline=True),
-                discord.EmbedField(name="AC: ", value=self.children[0].value, inline=True),
-                discord.EmbedField(name="Fort: ", value=self.children[1].value, inline=True),
-                discord.EmbedField(name="Reflex: ", value=self.children[2].value, inline=True),
-                discord.EmbedField(name="Will: ", value=self.children[3].value, inline=True),
-            ],
-            color=discord.Color.dark_gold(),
-        )
-
-        async_session = sessionmaker(self.engine, expire_on_commit=False, class_=AsyncSession)
-        Tracker = await get_tracker(self.ctx, self.engine, id=guild.id)
-
-        Condition = await get_condition(self.ctx, self.engine, id=guild.id)
-        async with async_session() as session:
-            char_result = await session.execute(select(Tracker).where(Tracker.name == self.name))
-            character = char_result.scalars().one()
-
-        for item in self.children:
-            async with async_session() as session:
-                result = await session.execute(
-                    select(Condition).where(Condition.character_id == character.id).where(Condition.title == item.label)
-                )
-                condition = result.scalars().one()
-                condition.number = int(item.value)
-                await session.commit()
-
-        Tracker_Model = await get_tracker_model(self.ctx, self.bot, guild=guild, engine=self.engine)
-        await Tracker_Model.update_pinned_tracker()
-        # print('Tracker Updated')
-        await self.ctx.channel.send(embeds=await Generic.character_functions.get_char_sheet(self.ctx, self.engine, self.bot, self.name))
-
-    async def on_error(self, error: Exception, interaction: Interaction) -> None:
-        logging.warning(error)
-        self.stop()
-
-
 async def D4eTrackerButtons(ctx: discord.ApplicationContext, bot, guild=None):
     engine = get_asyncio_db_engine(user=USERNAME, password=PASSWORD, host=HOSTNAME, port=PORT, db=SERVER_DATA)
     async_session = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
